dispatch rolling based on data, not prefix
In other words, the first table tells you what table to roll, but not
how to roll it.  This means you can direct to an encounter table
anywhere without caring what's in it.  It means that "goblin" could
start off as a "monster" entry but later be replaced with a standard
table to dispatch to different types.  It should make the whole thing
easier to plug together by turning tables into more standardized
components.