no more multi-doc files; "type" key
this was the original plan, but I decided to get clever.  stupid
rjbs, no clever!

this makes it much simpler to have embedded tables look just like
the contents of a file without having to worry about the stupidity
of pointless sequences

the rest of the type-and-header routine is there to guess at the
right type based on the contents of the struct