first lousy pass at fill in the blank tables
These are a mess because they do nothing to confirm that the final
result is something that will make sense inline.  We don't have any
sort of marker for "this result is a hunk of text with no newlines,"
for example.

It is the first set of tests that build nested tables from Perl data
structures, though.